+++ tests/testthat/test_cpos2lbound.R
@@ -16,6 +16,22 @@ test_that(
 )
 
 test_that(
+  "cpos_to_lbound",
+  {
+    cpos_old <- cl_cpos2lbound(
+      corpus = "REUTERS",
+      s_attribute = "places",
+      cpos = 5L,
+      registry = use_tmp_registry()
+    )
+    s <- s_attr(corpus = "REUTERS", s_attribute = "places", registry = get_tmp_registry())
+    cpos_new <- cpos_to_lbound(cpos = 5L, s_attr = s)
+    expect_identical(cpos_old, cpos_new)
+  }
+)
+
+
+test_that(
   "cpos2rbound",
   {
     cpos <- cl_cpos2rbound(
@@ -27,6 +43,22 @@ test_that(
     expect_equal(cpos, 91L)
   }
 )
+
+test_that(
+  "cpos_to_rbound",
+  {
+    cpos_old <- cl_cpos2rbound(
+      corpus = "REUTERS",
+      s_attribute = "places",
+      cpos = 5L,
+      registry = use_tmp_registry()
+    )
+    s <- s_attr(corpus = "REUTERS", s_attribute = "places", registry = get_tmp_registry())
+    cpos_new <- cpos_to_rbound(cpos = 5L, s_attr = s)
+    expect_identical(cpos_old, cpos_new)
+  }
+)
+
 
 test_that(
   "",